Move shared dependencies into neuro_san_studio package
