+++ coded_tools/experimental/mdap_decomposer/decomposition_solver.py
@@ -21,7 +21,7 @@ from neuro_san.interfaces.coded_tool import CodedTool
 from neuro_san.internals.graph.activations.branch_activation import BranchActivation
 
 from coded_tools.experimental.mdap_decomposer.neuro_san_solver import NeuroSanSolver
-from coded_tools.experimental.mdap_decomposer.solver_parsing import SolverParsing
+from neuro_san_studio.coded_tools.solver_parsing import SolverParsing
 from neuro_san_studio.coded_tools.coded_tool_agent_caller import CodedToolAgentCaller
