+++ neuro_san_studio/coded_tools/coded_tool_agent_caller.py
@@ -20,7 +20,7 @@ from typing import Any
 
 from neuro_san.internals.graph.activations.branch_activation import BranchActivation
 
-from coded_tools.experimental.mdap_decomposer.solver_parsing import SolverParsing
+from neuro_san_studio.coded_tools.solver_parsing import SolverParsing
 from neuro_san_studio.coded_tools.agent_caller import AgentCaller
